Add files via upload
This eBook will help you to master is the ability to solder electronics as this is the one of the most fundamental skills a maker should be able to do. Whether you are building a robot or working with Arduino or Raspberry Pi, knowing how to solder will come in handy. The best thing about soldering is that it is really fun to learn and easy to master.